getting the rest of the library back in line
