+++ include/universal/posit/math/sqrt.hpp
@@ -4,6 +4,7 @@
 // Copyright (C) 2017-2020 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
+#include "universal/native/ieee-754.hpp"
 #include "sqrt_tables.hpp"
 
 namespace sw {
